Remove unused option "exclude_useless_cards"
Usefulness of cards depends on situations.  So that it's not
a computable aspect.  Additionally, useless cards still have a power to
change flows of games by occuping stacks of the random market.
Therefore, this option is useless.

Since this option is not implemented and its UI is disabled,
it's safe to remove this option.

Close gh-47.